Create memoized component for the drag and drop list
