+++ src/components/BottomBar/SkyBrowser/SkyBrowserTabs.jsx
@@ -14,6 +14,7 @@ import SkyBrowserSettings from './SkyBrowserSettings.jsx'
 import {
   reloadPropertyTree,
 } from '../../../api/Actions';
+import SkyBrowserSelectedImagesList from './SkyBrowserSelectedImagesList';
 
 const ButtonIds = {
   LookAtTarget: "LookAtTarget",
@@ -49,7 +50,6 @@ function SkyBrowserTabs({
   );
   // Each message to WorldWide Telescope has a unique order number
   const [messageCounter, setMessageCounter] = React.useState(0);
-  const [isDragging, setIsDragging] = React.useState(false);
 
   // Cache for when Redux store is updating from OpenSpace
   // Using refs so the component doesn't update unnecessarily when they are set
@@ -414,77 +414,23 @@ function SkyBrowserTabs({
       return <></>;
     }
 
-    const onDragStart = () => {
-      setIsDragging(true);
-    };
-
-    const onDragEnd = async (result) => {
-      if (!result.destination || result.source.index === result.destination.index) {
-        setIsDragging(false);
-        return; // no change => do nothing
-      }
-
-      // First update the order manually, so we keep it while the properties
-      // are being refreshed below
-      // Images cache
-      const imagesOrder = isLoading.current ? [...imagesCache.current] : [...imageIndices];
-      imagesCache.current = getCurrentOrder(imagesOrder, result.source.index, result.destination.index);
-      // Opacities cache
-      const opacitiesOrder = isLoading.current ? [...opacitiesCache.current] : [...browsers[selectedBrowserId].opacities];
-      opacitiesCache.current = getCurrentOrder(opacitiesOrder, result.source.index, result.destination.index);
-
-      isLoading.current = true;
-      setIsDragging(false);
-
-      // Move image logic
-      await setImageLayerOrder(selectedBrowserId, Number(result.draggableId), result.destination.index)
-    };
-
-    // Invisible overlay that covers the entire body and prevents other hover effects
-    // from being triggered while dragging
-    const overlay = (
-      <div style={{
-        position: 'fixed', top: 0, right: 0, bottom: 0, left: 0, zIndex: 100,
-      }}
-      />
-    );
-
+  
     // Set image indices and opacities to the order they should currently have
     // Check if they are currently loading or not
     const imagesIndices = isLoading.current ? imagesCache.current : imageIndices;
     const images = imagesIndices.map(index => imageList[index.toString()]);
     const opacities = isLoading.current ? opacitiesCache.current : browsers[selectedBrowserId].opacities;
     return (
-      <DragDropContext onDragEnd={onDragEnd} onDragStart={onDragStart}>
-        { isDragging && overlay }
-        <Droppable droppableId="layers">
-          { provided => (
-            <div {...provided.droppableProps} ref={provided.innerRef}>
-              { images.map((entry, index) => (
-                <Draggable key={entry.identifier} draggableId={entry.identifier} index={index}>
-                  {provided => (
-                    <div {...provided.draggableProps} ref={provided.innerRef}>
-                      <SkyBrowserTabEntry
-                        dragHandleTitleProps={provided.dragHandleProps}
-                        {...entry}
-                        luaApi={luaApi}
-                        key={entry.identifier}
-                        onSelect={selectImage}
-                        removeImageSelection={removeImageSelection}
-                        opacity={opacities[index]}
-                        setOpacity={setOpacityOfImage}
-                        currentBrowserColor={currentBrowserColor}
-                        isActive={activeImage === entry.identifier}
-                      />
-                    </div>
-                  )}
-                </Draggable>
-              ))}
-              { provided.placeholder }
-            </div>
-          )}
-        </Droppable>
-      </DragDropContext>
+      <SkyBrowserSelectedImagesList
+        images={images}
+        opacities={opacities}
+        luaApi={luaApi}
+        selectImage={selectImage}
+        removeImageSelection={removeImageSelection}
+        setOpacityOfImage={setOpacityOfImage}
+        currentBrowserColor={currentBrowserColor}
+        activeImage={activeImage}
+      />
     );
   }
